feat: update build-nginx.sh to fetch latest PCRE version dynamically
This commit updates the build-nginx.sh script to fetch the latest version of PCRE dynamically from the GitHub API, instead of using a hardcoded version. It also adds 'jq' to the list of installed packages, which is used to parse the JSON response from the GitHub API. Additionally, the download URLs for the PCRE source files and their corresponding signature files have been updated to use the URLs provided in the GitHub API response.