feat(build-nginx.sh): update version lookup and nginx opgp key
This commit introduces two changes to the build-nginx.sh script:

Added a new line of code to look up the latest version of PCRE package from its source. This is done using a curl command to fetch the data from the source, and then using jq to parse the JSON response and extract the version information.

Updated the opgp key for nginx. The previous key was replaced with a new one. This is necessary to verify the authenticity of the downloaded nginx package.